Refactor system params into family modules
Move concrete params and structures into their owning system packages while keeping shared base params in systems.params. Extract GVS segment construction splitting into gvs.construction and update docs to describe the family-local layout. Fix API docstring warnings for tendon routing, HSA, and PCS constructors.
