+++ src/soromox/systems/articulated/__init__.py
@@ -1,4 +1,5 @@
+from soromox.systems.articulated.params import ArticulatedSoftRobotParams
+
 from .articulated_soft_robot import ArticulatedSoftRobot
-from soromox.systems.params import ArticulatedSoftRobotParams
 
 __all__ = ["ArticulatedSoftRobot", "ArticulatedSoftRobotParams"]
