+++ src/soromox/systems/gvs/specs.py
@@ -2,10 +2,10 @@ __all__ = ["GVSSegment", "LinkSpec", "JointSpec", "StrainBasisSpec"]
 
 from dataclasses import dataclass, field
 from typing import Literal
+
 from jax import Array
 
 from soromox.systems.soft_robot import CrossSectionGeometry
-
 
 StrainComponent = Literal[
     "kappa_x",
