+++ src/soromox/systems/hsa/__init__.py
@@ -1,5 +1,5 @@
-from soromox.systems.params import PlanarHSAParams
-from soromox.systems.structures import PlanarHSAStructure
+from soromox.systems.hsa.params import PlanarHSAParams
+from soromox.systems.hsa.structures import PlanarHSAStructure
 
 from .planar_hsa import PlanarHSA
