+++ src/soromox/systems/pcs/structures.py
@@ -1,0 +1,20 @@
+__all__ = ["PCSStructure", "PlanarPCSStructure"]
+
+import equinox as eqx
+from jax import Array
+
+
+class PCSStructure(eqx.Module):
+    """Static PCS layout that determines JAX compilation structure."""
+
+    num_gauss_points: int = eqx.field(static=True, default=5)
+    strain_selector: Array | None = None
+    scale_rotational_basis_by_length: bool = eqx.field(static=True, default=False)
+
+
+class PlanarPCSStructure(eqx.Module):
+    """Static planar PCS layout."""
+
+    num_gauss_points: int = eqx.field(static=True, default=5)
+    strain_selector: Array | None = None
+    scale_rotational_basis_by_length: bool = eqx.field(static=True, default=False)
