+++ src/soromox/systems/pendulum/__init__.py
@@ -1,6 +1,10 @@
+from soromox.systems.pendulum.params import (
+    PendulumParams,
+    TendonActuatedPendulumParams,
+)
+
 from .pendulum import Pendulum
 from .tendon_actuated_pendulum import TendonActuatedPendulum
-from soromox.systems.params import PendulumParams, TendonActuatedPendulumParams
 
 __all__ = [
     "Pendulum",
